p2p/secret_connection: Combine HKDF calls
This commit merges the three calls to hkdf into a single call.
Notably the hkdf call is now just called upon the computed shared
secret, since the shared secret is a function of the pubkeys.

It also makes nonces start at 0, as we are using chacha as a stream
cipher.